Refactored input loop service
