+++ samples/ravendb/simple/Raven_9/Client/Program.cs
@@ -1,5 +1,4 @@
 using System;
-using Client;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using NServiceBus;
@@ -7,7 +6,6 @@ using NServiceBus;
 Console.Title = "Client";
 
 var builder = Host.CreateApplicationBuilder(args);
-builder.Services.AddHostedService<InputLoopService>();
 
 var endpointConfiguration = new EndpointConfiguration("Samples.RavenDB.Client");
 
@@ -16,4 +14,29 @@ endpointConfiguration.UseTransport(new LearningTransport());
 
 Console.WriteLine("Press 'enter' to send a StartOrder messages");
 builder.UseNServiceBus(endpointConfiguration);
-await builder.Build().RunAsync();
+
+var host = builder.Build();
+await host.StartAsync();
+
+var messageSession = host.Services.GetRequiredService<IMessageSession>();
+
+while (true)
+{
+    var key = Console.ReadKey();
+    Console.WriteLine();
+
+    if (key.Key != ConsoleKey.Enter)
+    {
+        break;
+    }
+
+    var orderId = Guid.NewGuid();
+    var startOrder = new StartOrder
+    {
+        OrderId = orderId
+    };
+    await messageSession.Send("Samples.RavenDB.Server", startOrder);
+    Console.WriteLine($"StartOrder Message sent with OrderId {orderId}");
+}
+
+await host.StopAsync();
