--- samples/ravendb/simple/Raven_9/Server/BackgroundService.cs
@@ -1,6 +1,0 @@
-﻿namespace Server
-{
-    public class BackgroundService
-    {
-    }
-}
