+++ samples/ravendb/simple/Raven_9/Server/Program.cs
@@ -40,10 +40,6 @@ endpointConfiguration.UseTransport(transport);
 
 await EnsureDatabaseExists(documentStore);
 
-
-
-Console.WriteLine("Press any key, the application is starting");
-Console.ReadKey();
 Console.WriteLine("Starting...");
 
 builder.UseNServiceBus(endpointConfiguration);
